+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec
 - Power sum explicit view
 - Faulhaber's coefficients mathematica
 - Rearrange subsection folders
+- Multifold decompositions
+- Multifold decompositions mathematica
 
 ## [1.0.2] - 2025-07-27